pyproject.toml
[tool.poetry]
name = "find_rogues"
version = "0.1.0"
description = ""
authors = ["Michael Rose Jr. <michael@michaelrosejr.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycentral = "^0.0.3"
rich = "^12.6.0"
PyYAML = "^6.0"
pendulum = "^2.1.2"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.20.8"
sendgrid = "^6.9.7"
Jinja2 = "^3.1.2"
typer = "^0.7.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
